requirements.txt
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
python-dotenv
streamlit
unstructured